refactor(memory): extract setup.rs test module to sibling file (line-cap, #914 PR2)
Move the inline `#[cfg(test)] mod tests { … }` block out of
`crates/trusty-memory/src/commands/setup.rs` into a sibling file
`setup_tests.rs`, referenced via `#[path = "setup_tests.rs"] mod tests`.

Before: setup.rs = 784 lines (1 over the frozen allowlist budget of 783).
After:  setup.rs = 539 lines, setup_tests.rs = 259 lines (both ≤ 500 cap goal missed
        by 39 lines for setup.rs, but allowlist budget drops 783 → 539).

All 6 moved tests (patch_one_creates_missing_file, patch_one_is_idempotent,
patch_one_installs_hook_when_mcp_already_present,
patch_one_installs_session_start_hook_when_upgrading,
patch_one_adds_session_start_when_legacy_user_prompt_submit_has_different_shape,
patch_one_preserves_unrelated_keys) pass unchanged.
